agentarea-bootstrap/pyproject.toml
[project]
name = "agentarea-bootstrap"
version = "0.0.1"
description = "Bootstrap module for AgentArea - sets up initial configuration, LLM providers, MinIO, and Infisical"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
    "psycopg2-binary>=2.9.9",
    "pyyaml>=6.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
